Move tests not specific to FB into a mixin
